applications/payment-service/requirements.txt
flask==3.1.*
gunicorn==23.0.*
opentelemetry-api==1.33.*
opentelemetry-sdk==1.33.*
opentelemetry-exporter-otlp-proto-grpc==1.33.*
opentelemetry-instrumentation-flask==0.54b0
opentelemetry-instrumentation-logging==0.54b0
python-json-logger==3.2.*
